finished m09

+++ module09/ex02/main.cpp
@@ -6,7 +6,7 @@
 /*   By: mademir <mademir@student.codam.nl>           +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2024/03/04 13:47:18 by mademir       #+#    #+#                 */
-/*   Updated: 2024/04/08 11:21:28 by mademir       ########   odam.nl         */
+/*   Updated: 2024/04/12 10:59:54 by mademir       ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -51,7 +51,7 @@ void	iter_swap(PmergeMe<Container> lhs, PmergeMe<Container> rhs)
     std::swap_ranges(lhs.getIter(), lhs.getIter() + lhs.getSize(), rhs.getIter());
 }
 
-// Grouping function
+// Grouping functions
 template<typename Container>
 PmergeMe<Container>	makeGroupIter(Container it, std::size_t size)
 {
@@ -81,14 +81,14 @@ void	fordJohnson(Iterator first, Iterator last)
     }
 
     fordJohnson(makeGroupIter(first, 2), makeGroupIter(end, 2));
-	
+
     // Separate main chain and pend elements
     std::list<Iterator> mainChain;
 	mainChain.push_back(first);
 	mainChain.push_back(std::next(first));
 
     std::list<typename std::list<Iterator>::iterator> pend;
-    for (Iterator it = first + 2 ; it != end ; it += 2)
+    for (Iterator it = first + 2; it != end; it += 2)
     {
         typename std::list<Iterator>::iterator tmp = mainChain.insert(std::end(mainChain), std::next(it));
         pend.push_back(tmp);
@@ -106,7 +106,8 @@ void	fordJohnson(Iterator first, Iterator last)
 	for (int k = 0; ; ++k)
 	{
 		std::size_t dist = jacobsthal_diff[k];
-		if (dist > static_cast<std::size_t>(std::distance(current_pend, std::end(pend)))) break;
+		if (dist > static_cast<std::size_t>(std::distance(current_pend, std::end(pend))))
+			break;
 
 		Iterator it = std::next(current_it, dist * 2);
 		typename std::list<typename std::list<Iterator>::iterator>::iterator pe = std::next(current_pend, dist);
@@ -130,8 +131,7 @@ void	fordJohnson(Iterator first, Iterator last)
         current_it += 2;
         ++current_pend;
     }
-		
-    // Move values in order to cache, then back
+
     std::list<typename Iterator::value_type> cache;
     for (Iterator& it: mainChain)
     {
@@ -147,13 +147,13 @@ int	main(int argc, char **argv)
 	if (argc < 2)
 		exitError("Usage: ./PmergeMe <sequence of positive integers>");
 
-	std::vector<int>	vec;
-	std::deque<int>		dq;
-	
+	std::vector<long>	vec;
+	std::deque<long>	dq;
+
 	for (int i = 1; i < argc; i++)
 	{
 		std::istringstream ss(argv[i]);
-		int num;
+		long num;
 		if (!(ss >> num) || !ss.eof())
 			exitError("Error: input invalid");
 		if (num < 0)
